+++ crates/railgun/src/circuit/mod.rs
@@ -1,5 +1,0 @@
-pub mod groth16_prover;
-pub mod inputs;
-pub mod proof;
-mod remote_artifact_loader;
-mod witness;