+++ crates/railgun/src/indexer/mod.rs
@@ -1,4 +1,0 @@
-pub(crate) mod indexed_account;
-pub mod syncer;
-pub(crate) mod txid_indexer;
-pub(crate) mod utxo_indexer;