+++ crates/tornadocash/src/provider/mod.rs
@@ -1,4 +1,0 @@
-pub mod note;
-pub mod pool;
-pub mod pool_provider;
-pub mod provider;