+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @kohaku-eth/plugins
 
+## 0.0.1-alpha.14
+
+### Patch Changes
+
+- Updated dependencies [e3735d4]
+  - @kohaku-eth/provider@0.1.0-alpha.10
+
+## 0.0.1-alpha.13
+
+### Patch Changes
+
+- Updated dependencies [8d5a29e]
+  - @kohaku-eth/provider@0.1.0-alpha.9
+
 ## 0.0.1-alpha.12
 
 ### Patch Changes