+++ packages/tornado-cash/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.0.2-alpha.22
+
+### Patch Changes
+
+- Updated dependencies [e3735d4]
+  - @kohaku-eth/provider@0.1.0-alpha.10
+  - @kohaku-eth/plugins@0.0.1-alpha.14
+
+## 0.0.2-alpha.21
+
+### Patch Changes
+
+- Updated dependencies [8d5a29e]
+  - @kohaku-eth/provider@0.1.0-alpha.9
+  - @kohaku-eth/plugins@0.0.1-alpha.13
+
 ## 0.0.2-alpha.20
 
 ### Patch Changes